requirements.txt
undetected_chromedriver==3.5.5
bs4==0.0.2
requests==2.26.0
lxml==4.9.3
selenium==4.23.0
openai==1.12.0
google-generativeai==0.3.2
schedule==1.2.1
python-dotenv==1.0.0
tqdm==4.66.1
pyyaml==6.0.1